Removed every single warning when compiling with -Wall on Ubuntu from the source base. Cleans up the code a bit.
git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@451 5e24083e-1465-11df-9d2a-2365f6df44e6

+++ Engine/src/AnimData.cpp
@@ -183,7 +183,7 @@ bool AnimData::ParseStrip(
         anim.frameTimes.push_back( time );
       }
       if ( anim.frameTimes.size() > 1 ) {
-        if ( anim.frameTimes.size() == anim.numFrames ) {
+        if ( static_cast<int>(anim.frameTimes.size()) == anim.numFrames ) {
           anim.uniqueFrameTimes = true;
         } else {
           LogErr( "When specifying multiple frame times, didn't specify time "
